Always retry `contains partially-complete state` error if we have a `receive_resume_token`
Previously we only retried this error if we aborted a stuck replication ourselves.
We should also retry it if the replication was aborted by 3rd party (e.g. network error)
